requirements.txt
tqdm==4.67.0
matplotlib==3.9.2
datasets==3.1.0
pydantic==2.9.2
click==8.1.7
kornia==0.7.4
einops==0.8.0
einops-exts==0.0.4
beartype==0.19.0
transformers==4.46.3
pytorch-warmup==0.1.1
ema-pytorch==0.7.3
accelerate==1.1.1
tensorboard==2.19.0
scipy==1.15.3
opencv-python==4.11.0.86